+++ .cache/0a/9e34320923e65f03576121293fc79d.json
@@ -1,0 +1,1 @@
+{"id":"../node_modules/remove-trailing-slash/index.js","dependencies":[{"name":"/home/ray/Desktop/Router/package.json","includedInParent":true,"mtime":1595186753722},{"name":"/home/ray/Desktop/Router/node_modules/remove-trailing-slash/package.json","includedInParent":true,"mtime":1595105733690}],"generated":{"js":"\n/**\n * Remove trailing slashes from the given `str`\n *\n * @api public\n * @param {String} str\n * @return {String}\n */\nexports = module.exports = function (str) {\n  return String(str).replace(exports.expr, '');\n};\n\n/**\n * Expose the regex\n *\n * @api private\n * @type {RegExp}\n */\nexports.expr = /\\/+$/;\n"},"sourceMaps":{"js":{"mappings":[{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":2,"column":0},"generated":{"line":2,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":3,"column":0},"generated":{"line":3,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":4,"column":0},"generated":{"line":4,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":5,"column":0},"generated":{"line":5,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":6,"column":0},"generated":{"line":6,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":7,"column":0},"generated":{"line":7,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":8,"column":0},"generated":{"line":8,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":9,"column":0},"generated":{"line":9,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":10,"column":0},"generated":{"line":10,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":11,"column":0},"generated":{"line":11,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":12,"column":0},"generated":{"line":12,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":13,"column":0},"generated":{"line":13,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":14,"column":0},"generated":{"line":14,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":15,"column":0},"generated":{"line":15,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":16,"column":0},"generated":{"line":16,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":17,"column":0},"generated":{"line":17,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":18,"column":0},"generated":{"line":18,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":19,"column":0},"generated":{"line":19,"column":0}},{"source":"../node_modules/remove-trailing-slash/index.js","original":{"line":20,"column":0},"generated":{"line":20,"column":0}}],"sources":{"../node_modules/remove-trailing-slash/index.js":"\n/**\n * Remove trailing slashes from the given `str`\n *\n * @api public\n * @param {String} str\n * @return {String}\n */\nexports = module.exports = function (str) {\n  return String(str).replace(exports.expr, '');\n};\n\n/**\n * Expose the regex\n *\n * @api private\n * @type {RegExp}\n */\nexports.expr = /\\/+$/;\n"},"lineCount":20}},"error":null,"hash":"a74168a20f01b776460345c9aa1f8002","cacheData":{"env":{}}}
