+++ .cache/47/9d7d62416f18bf0ea845ef368ee9e4.json
@@ -1,0 +1,1 @@
+{"id":"../src/parser/index.ts","dependencies":[{"name":"/home/ray/Desktop/Router/package.json","includedInParent":true,"mtime":1595186753722},{"name":"./parser","loc":{"line":13,"column":21},"parent":"/home/ray/Desktop/Router/src/parser/index.ts","resolved":"/home/ray/Desktop/Router/src/parser/parser.ts"}],"generated":{"js":"\"use strict\";\n\nvar __createBinding = this && this.__createBinding || (Object.create ? function (o, m, k, k2) {\n  if (k2 === undefined) k2 = k;\n  Object.defineProperty(o, k2, {\n    enumerable: true,\n    get: function get() {\n      return m[k];\n    }\n  });\n} : function (o, m, k, k2) {\n  if (k2 === undefined) k2 = k;\n  o[k2] = m[k];\n});\n\nvar __exportStar = this && this.__exportStar || function (m, exports) {\n  for (var p in m) {\n    if (p !== \"default\" && !exports.hasOwnProperty(p)) __createBinding(exports, m, p);\n  }\n};\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\n__exportStar(require(\"./parser\"), exports);"},"sourceMaps":{"js":{"mappings":[{"source":"../src/parser/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":26,"column":0}},{"source":"../src/parser/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":26,"column":12}},{"source":"../src/parser/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":26,"column":13}},{"source":"../src/parser/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":26,"column":20}},{"source":"../src/parser/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":26,"column":21}},{"source":"../src/parser/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":26,"column":31}},{"source":"../src/parser/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":26,"column":32}},{"source":"../src/parser/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":26,"column":34}},{"source":"../src/parser/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":26,"column":41}},{"source":"../src/parser/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":26,"column":42}}],"sources":{"../src/parser/index.ts":"export * from \"./parser\";\n"},"lineCount":null}},"error":null,"hash":"94afe5ba50f6b8d43a5e281994ad98d8","cacheData":{"env":{}}}
