+++ .cache/61/ca0171a67771536c86c59a52386136.json
@@ -1,0 +1,1 @@
+{"id":"../src/utils/trimSlash.ts","dependencies":[{"name":"/home/ray/Desktop/Router/package.json","includedInParent":true,"mtime":1595186753722},{"name":"remove-trailing-slash","loc":{"line":7,"column":54},"parent":"/home/ray/Desktop/Router/src/utils/trimSlash.ts","resolved":"/home/ray/Desktop/Router/node_modules/remove-trailing-slash/index.js"}],"generated":{"js":"\"use strict\";\n\nvar __importDefault = this && this.__importDefault || function (mod) {\n  return mod && mod.__esModule ? mod : {\n    \"default\": mod\n  };\n};\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.trimTrailingSlash = void 0;\n\nvar remove_trailing_slash_1 = __importDefault(require(\"remove-trailing-slash\"));\n\nvar trimTrailingSlash = function trimTrailingSlash(path) {\n  return remove_trailing_slash_1.default(path);\n};\n\nexports.trimTrailingSlash = trimTrailingSlash;"},"sourceMaps":{"js":{"mappings":[{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":14,"column":0}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":14,"column":4}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":14,"column":27}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":14,"column":30}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":14,"column":45}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":14,"column":46}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":14,"column":53}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":14,"column":54}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":14,"column":77}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":14,"column":78}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":14,"column":79}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":0},"generated":{"line":16,"column":0}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":6},"generated":{"line":16,"column":4}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":23},"generated":{"line":16,"column":21}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":26},"generated":{"line":16,"column":24}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":6},"generated":{"line":16,"column":33}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":26},"generated":{"line":16,"column":50}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":27},"generated":{"line":16,"column":51}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":26},"generated":{"line":16,"column":55}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":39},"generated":{"line":16,"column":57}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":52},"generated":{"line":17,"column":0}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":52},"generated":{"line":17,"column":9}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":52},"generated":{"line":17,"column":32}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":52},"generated":{"line":17,"column":33}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":52},"generated":{"line":17,"column":40}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":72},"generated":{"line":17,"column":41}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":52},"generated":{"line":17,"column":45}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":52},"generated":{"line":17,"column":46}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":77},"generated":{"line":18,"column":0}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":3,"column":0},"generated":{"line":18,"column":1}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":20,"column":0}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":20,"column":7}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":20,"column":8}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":20,"column":25}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":20,"column":28}},{"source":"../src/utils/trimSlash.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":20,"column":45}}],"sources":{"../src/utils/trimSlash.ts":"import removeTrailingSlash from \"remove-trailing-slash\";\n\nconst trimTrailingSlash = (path: string): string => removeTrailingSlash(path);\n\nexport { trimTrailingSlash };\n"},"lineCount":null}},"error":null,"hash":"33bcda465c2d8b92c2883fd20b9d256e","cacheData":{"env":{}}}
