+++ .cache/7f/ed53cc024bf57251864957131a24b3.json
@@ -1,0 +1,1 @@
+{"id":"../src/utils/index.ts","dependencies":[{"name":"/home/ray/Desktop/Router/package.json","includedInParent":true,"mtime":1595186753722},{"name":"./trimSlash","loc":{"line":3,"column":26},"parent":"/home/ray/Desktop/Router/src/utils/index.ts","resolved":"/home/ray/Desktop/Router/src/utils/trimSlash.ts"}],"generated":{"js":"\"use strict\";\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar trimSlash_1 = require(\"./trimSlash\");\n\nObject.defineProperty(exports, \"trimTrailingSlash\", {\n  enumerable: true,\n  get: function get() {\n    return trimSlash_1.trimTrailingSlash;\n  }\n});"},"sourceMaps":{"js":{"mappings":[{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":7,"column":0}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":7,"column":4}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":7,"column":15}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":7,"column":18}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":7,"column":25}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":7,"column":26}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":7,"column":39}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":7,"column":40}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":9,"column":0}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":9,"column":6}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":9,"column":7}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":9,"column":21}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":9,"column":22}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":9,"column":29}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":9,"column":31}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":9,"column":50}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":9,"column":52}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":10,"column":0}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":10,"column":2}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":10,"column":12}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":10,"column":14}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":10,"column":18}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":11,"column":0}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":11,"column":2}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":11,"column":5}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":11,"column":7}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":11,"column":22}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":12,"column":0}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":12,"column":11}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":12,"column":22}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":12,"column":23}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":12,"column":40}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":26},"generated":{"line":13,"column":0}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":14,"column":0}},{"source":"../src/utils/index.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":14,"column":1}}],"sources":{"../src/utils/index.ts":"export { trimTrailingSlash } from \"./trimSlash\";\n"},"lineCount":null}},"error":null,"hash":"1f70a06f18e57b95490833187c64336d","cacheData":{"env":{}}}
