+++ .cache/dc/47ebb457a9044fe7fa2c91d504e06b.json
@@ -1,0 +1,1 @@
+{"id":"../src/communication/request.ts","dependencies":[{"name":"/home/ray/Desktop/Router/package.json","includedInParent":true,"mtime":1595186753722}],"generated":{"js":"\"use strict\";\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.RouterRequest = void 0;\n\nvar RouterRequest =\n/** @class */\nfunction () {\n  function RouterRequest() {}\n\n  RouterRequest.prototype.getPath = function () {\n    return window.location.pathname;\n  };\n\n  RouterRequest.prototype.onUrlChange = function (callback) {};\n\n  return RouterRequest;\n}();\n\nexports.RouterRequest = RouterRequest;"},"sourceMaps":{"js":{"mappings":[{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":8,"column":0}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":8,"column":4}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":8,"column":17}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":9,"column":0}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":10,"column":0}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":10,"column":12}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":11,"column":0}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":11,"column":11}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":11,"column":24}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":11,"column":27}},{"source":"../src/communication/request.ts","name":null,"original":{"line":6,"column":1},"generated":{"line":11,"column":28}},{"source":"../src/communication/request.ts","name":null,"original":{"line":2,"column":9},"generated":{"line":13,"column":0}},{"source":"../src/communication/request.ts","name":null,"original":{"line":2,"column":9},"generated":{"line":13,"column":2}},{"source":"../src/communication/request.ts","name":null,"original":{"line":2,"column":9},"generated":{"line":13,"column":15}},{"source":"../src/communication/request.ts","name":null,"original":{"line":2,"column":9},"generated":{"line":13,"column":16}},{"source":"../src/communication/request.ts","name":null,"original":{"line":2,"column":9},"generated":{"line":13,"column":25}},{"source":"../src/communication/request.ts","name":null,"original":{"line":2,"column":9},"generated":{"line":13,"column":26}},{"source":"../src/communication/request.ts","name":null,"original":{"line":2,"column":9},"generated":{"line":13,"column":33}},{"source":"../src/communication/request.ts","name":null,"original":{"line":2,"column":2},"generated":{"line":13,"column":36}},{"source":"../src/communication/request.ts","name":null,"original":{"line":2,"column":2},"generated":{"line":13,"column":48}},{"source":"../src/communication/request.ts","name":null,"original":{"line":3,"column":4},"generated":{"line":14,"column":0}},{"source":"../src/communication/request.ts","name":null,"original":{"line":3,"column":11},"generated":{"line":14,"column":11}},{"source":"../src/communication/request.ts","name":null,"original":{"line":3,"column":17},"generated":{"line":14,"column":17}},{"source":"../src/communication/request.ts","name":null,"original":{"line":3,"column":18},"generated":{"line":14,"column":18}},{"source":"../src/communication/request.ts","name":null,"original":{"line":3,"column":11},"generated":{"line":14,"column":26}},{"source":"../src/communication/request.ts","name":null,"original":{"line":3,"column":27},"generated":{"line":14,"column":27}},{"source":"../src/communication/request.ts","name":null,"original":{"line":3,"column":4},"generated":{"line":14,"column":35}},{"source":"../src/communication/request.ts","name":null,"original":{"line":4,"column":3},"generated":{"line":15,"column":0}},{"source":"../src/communication/request.ts","name":null,"original":{"line":2,"column":9},"generated":{"line":15,"column":3}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":17,"column":0}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":17,"column":2}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":17,"column":15}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":17,"column":16}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":17,"column":25}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":17,"column":26}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":17,"column":37}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":2},"generated":{"line":17,"column":40}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":21},"generated":{"line":17,"column":50}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":2},"generated":{"line":17,"column":58}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":39},"generated":{"line":17,"column":60}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":43},"generated":{"line":17,"column":61}},{"source":"../src/communication/request.ts","name":null,"original":{"line":5,"column":9},"generated":{"line":17,"column":62}},{"source":"../src/communication/request.ts","name":null,"original":{"line":6,"column":0},"generated":{"line":19,"column":0}},{"source":"../src/communication/request.ts","name":null,"original":{"line":6,"column":0},"generated":{"line":19,"column":9}},{"source":"../src/communication/request.ts","name":null,"original":{"line":6,"column":0},"generated":{"line":19,"column":22}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":20,"column":0}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":20,"column":1}},{"source":"../src/communication/request.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":20,"column":3}},{"source":"../src/communication/request.ts","name":null,"original":{"line":7,"column":9},"generated":{"line":22,"column":0}},{"source":"../src/communication/request.ts","name":null,"original":{"line":7,"column":9},"generated":{"line":22,"column":7}},{"source":"../src/communication/request.ts","name":null,"original":{"line":7,"column":9},"generated":{"line":22,"column":8}},{"source":"../src/communication/request.ts","name":null,"original":{"line":7,"column":9},"generated":{"line":22,"column":21}},{"source":"../src/communication/request.ts","name":null,"original":{"line":7,"column":9},"generated":{"line":22,"column":24}},{"source":"../src/communication/request.ts","name":null,"original":{"line":7,"column":9},"generated":{"line":22,"column":37}}],"sources":{"../src/communication/request.ts":"class RouterRequest {\n  public getPath(): string {\n    return window.location.pathname;\n  }\n  public onUrlChange(callback: Function) {}\n}\nexport { RouterRequest };\n"},"lineCount":null}},"error":null,"hash":"d95893c2d9f6ffdb6871ca562888a1c3","cacheData":{"env":{}}}
