+++ src/communication/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from "./router-request-interface";
+export * from "./router-response-interface";
