+++ src/communication/interfaces/router-request-interface.ts
@@ -1,0 +1,5 @@
+interface RouterRequestInterface {
+  getPath(): string;
+  onRequest(callback): void;
+}
+export { RouterRequestInterface };
