+++ src/communication/interfaces/router-response-interface.ts
@@ -1,0 +1,4 @@
+interface RouterResponseInterface {
+  redirect(path: string): void;
+}
+export { RouterResponseInterface };
