--- src/communication/readme.md
@@ -1,1 +1,0 @@
-Navigation related functionality package
