+++ src/communication/response.ts
@@ -1,4 +1,5 @@
-class RouterResponse {
+import { RouterResponseInterface } from "./interfaces";
+class RouterResponse implements RouterResponseInterface {
   public redirect(path: string) {
     history.pushState({}, "", path);
     window.dispatchEvent(new Event("popstate"));
