--- src/core/readme.md
@@ -1,1 +1,0 @@
-Router core functionality package
