--- src/parser/readme.md
@@ -1,1 +1,0 @@
-Path parsing related functionality package
